refactor(tests): move unit tests into similar package as task classes
